src/requirements.txt
agno==1.7.0
groq==0.29.0
psycopg2-binary==2.9.10
psycopg==3.2.9
sqlalchemy==2.0.41
pydantic==2.11.7
uvicorn==0.35.0
fastapi==0.115.14
python-dotenv==1.1.0
pypdf==5.7.0
streamlit==1.46.1
faiss-cpu==1.11.0
sentence-transformers==5.0.0
langchain==0.3.26
langchain-community==0.3.27
